refactor: rename internal modules
