+++ covsirphy/gis/_choropleth.py
